refactor: Move copy button to separate button
