+++ src/course/course-code-block.component.tsx
@@ -1,12 +1,9 @@
-import { useEffect, useRef, useState } from 'react';
-import { Check, Copy } from 'lucide-react';
+import { useEffect, useState } from 'react';
 
 import { cn } from '@/shadcn/lib/utils';
+import { CourseCodeBlockCopyBtn } from './course-code-block-copy-btn.component';
 import { highlightCourseCode } from './course-code-block.shiki';
-import type {
-  CopyState,
-  CourseCodeBlockProps,
-} from './course-code-block.types';
+import type { CourseCodeBlockProps } from './course-code-block.types';
 
 const codeBlockClassName = cn(
   'course-code-block my-4 overflow-hidden rounded-2xl',
@@ -24,16 +21,6 @@ const languageBadgeClassName = cn(
   'shrink-0 rounded-full px-2 py-0.5 uppercase tracking-wide',
   'bg-black/5 text-slate-600',
   'dark:bg-white/10 dark:text-zinc-300'
-);
-
-const copyButtonClassName = cn(
-  'inline-flex shrink-0 cursor-pointer items-center gap-1.5',
-  'rounded-lg border px-2.5 py-1 text-xs font-medium',
-  'border-black/10 bg-white text-slate-700 transition-colors',
-  'hover:bg-slate-100 focus-visible:outline-none',
-  'focus-visible:ring-2 focus-visible:ring-slate-400/40',
-  'dark:border-white/10 dark:bg-white/5 dark:text-zinc-200',
-  'dark:hover:bg-white/10 dark:focus-visible:ring-white/40'
 );
 
 const highlightedCodeClassName = cn(
@@ -88,36 +75,8 @@ export function CourseCodeBlock({
   language,
   fileName,
 }: CourseCodeBlockProps) {
-  const [copyState, setCopyState] = useState<CopyState>('idle');
-  const copyResetTimeoutRef = useRef<number | null>(null);
   const displayFileName = fileName ?? 'Фрагмент кода';
   const codeBodyKey = `${language ?? 'plain'}\u0000${code}`;
-
-  useEffect(() => {
-    return () => {
-      if (copyResetTimeoutRef.current !== null) {
-        window.clearTimeout(copyResetTimeoutRef.current);
-      }
-    };
-  }, []);
-
-  async function handleCopy() {
-    if (copyResetTimeoutRef.current !== null) {
-      window.clearTimeout(copyResetTimeoutRef.current);
-    }
-
-    try {
-      await navigator.clipboard.writeText(code);
-      setCopyState('copied');
-    } catch {
-      setCopyState('failed');
-    }
-
-    copyResetTimeoutRef.current = window.setTimeout(() => {
-      setCopyState('idle');
-      copyResetTimeoutRef.current = null;
-    }, 1400);
-  }
 
   return (
     <figure className={codeBlockClassName}>
@@ -129,23 +88,7 @@ export function CourseCodeBlock({
           <span className="truncate">{displayFileName}</span>
         </div>
 
-        <button
-          type="button"
-          onClick={handleCopy}
-          className={copyButtonClassName}
-          aria-label="Скопировать код"
-        >
-          {copyState === 'copied' ? (
-            <Check className="size-3.5" aria-hidden="true" />
-          ) : (
-            <Copy className="size-3.5" aria-hidden="true" />
-          )}
-          {copyState === 'copied'
-            ? 'Скопировано'
-            : copyState === 'failed'
-              ? 'Не удалось'
-              : 'Скопировать'}
-        </button>
+        <CourseCodeBlockCopyBtn code={code} />
       </figcaption>
 
       <CourseCodeBlockBody key={codeBodyKey} code={code} language={language} />
