+++ src/course/course-code-block.types.ts
@@ -3,5 +3,3 @@ export type CourseCodeBlockProps = {
   language?: string;
   fileName?: string;
 };
-
-export type CopyState = 'idle' | 'copied' | 'failed';
